+++ codegen/src/rust.rs
@@ -1058,7 +1058,8 @@ impl<'a> RustGenerator<'a> {
         genln!(self, "    pub const VERSION: u32 = {};", svc.version().value());
         genln!(self);
         genln!(self, "    pub async fn new(object: &aldrin::Object) -> Result<Self, aldrin::Error> {{");
-        genln!(self, "        let inner = object.create_service(Self::UUID, Self::VERSION).await?;");
+        genln!(self, "        let info = aldrin::core::ServiceInfo::new(Self::VERSION);");
+        genln!(self, "        let inner = object.create_service(Self::UUID, info).await?;");
         genln!(self, "        Ok({} {{ inner }})", svc_name);
         genln!(self, "    }}");
         genln!(self);